Restructure into single package
