+++ src/orcasound_noise/utils/__init__.py
@@ -1,0 +1,2 @@
+from .hydrophone import Hydrophone
+from .file_connector import S3FileConnector
